Update README.md
This is the pre-release of the app behind https://t.me/CinemaCityAppBot bot. 

CinemaCityAppBot is the only mobile movie ticketing app that lets you buy movie tickets, make movie plans with friends & get a VIP experience at the theater so you never wait in line. When you get to the movie theater, skip the box office line & scan your digital movie ticket with your phone.

Skip the line and watch new movies as soon as they hit theaters with the Cinema City Telegram App.

Of course, this is a fake cinema, and the app is solely built to demonstrate some of Telegram's craziest features.

Good luck!